[US_AZ] Update source of sentence expiration date (Recidiviz/recidiviz-data#34288)
## Description of the change

A few more learnings about AZ sentences, and a lot of validation against
their OMS, made it clear that we were pulling the sentence expiration
date (`projected_release_date_max_external`) from the wrong source field
in raw data. This changes the `state_sentence_group` view to pull this
date from the `AZ_DOC_SC_OFFENSE` raw data table, which is where the
rest of the dates for sentence group lengths come from.

I validated a number of cases against ACIS (the AZ OMS) to make sure
that this is the field that is displayed as a resident's Sentence
Expiration Date in the frontend.

Despite only touching the `projected_release_date_max_external` (SED)
value, this change will help increase accuracy of our
`projected_release_date_min_external` (ERCD) value. Prior to this
change, the incorrect SED was sometimes before the ERCD value;
normalization auto-sorted these dates so that the smaller one would be
stored in the `min` field and the larger one stored in the `max` field.
This was causing incorrect SEDs to appear as ERCDs. Now that we have the
proper SEDs, this unexpected sorting will not happen, and we will see
the correct ERCD dates that we were already ingesting in the
`projected_release_date_min_external` field.

Tests:
- [x] ingest view query is deterministic
- [x] run mappings script passes
- [x] sandbox ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-10-17_13_02_06-14708763701472606667;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
succeeds

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 53ced209a07b066b5159d6fa53466887341938ff