[Justice Counts] Check if agency has county before getting county code and county name (Recidiviz/recidiviz-data#34218)
## Description of the change

An
[error](https://recidiviz-inc.sentry.io/issues/5996731356/?project=4504532096516096&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&stream_index=0)
came up in Sentry when someone tried to use the dashboard API for an
agency that did not have a county. This PR adds logic to check if an
agency has a county before trying to compute county code and name.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0abfb7c06d9b930c770fcfb06cfa1168a2a2f145