[US_AZ] Ingest ACCAT data into state_assessment (Recidiviz/recidiviz-data#34313)
## Description of the change

This PR ingests the total ACCAT score and the final supervision level
associated with that score into `state_assessment`.

ACCAT is short for Arizona Community Corrections Assessment Tool. This
is one of a handful of assessment tools used by ADCRR; it is the one
that is used currently to determine clients' supervision levels. There
is possible future work to ingest other assessment types, but for the
time being this is the most helpful assessment to have ingested for LRM
and policy engineering work.

Sandbox ingest pipeline succeeded
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-10-18_12_21_25-74542768869013610;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34311

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 272f5e90036aaf015b3d75fc9b38714f4b58f11e