[US_AZ] Improve state task deadline ingest logic (Recidiviz/recidiviz-data#34308)
## Description of the change

We are using the state_task_deadline entity in AZ to identify two
release dates stored in their OMS (ACIS TPR and DTP dates). The last
iteration of the view that hydrates this entity made a few mistakes that
are corrected here:
1. We should be partitioning by FINAL_OFFENSE_ID (sentence) instead of
DOC_ID (person incarceration stint). The results of these should
theoretically be the same, but since these release dates are calculated
at the sentence group level, it is more logically sound to orient this
view around sentences.
2. There are more ways to deduplicate changes that were made with the
same UPDT_DTM than were originally factored into this view. This PR adds
deduplication using CHANGE_ID and SC_CALC_HISTORY_ID, which are two
identifiers that only increment over time. This limits the number of
times we need to deduplicate by the actual eligibility date to 7.
3. In those 7 cases, we preserve the latest eligibility date rather than
the earliest one. This is more aligned with reality based on what we
have learned about how these dates change over time.

Tests:
- [x] ingest view query is deterministic
- [x] run_mappings script passes
- [x] sandbox ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-10-18_10_16_18-4041155279008777014;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
succeeds
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29234
Related to Recidiviz/recidiviz-data#34237

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 56593d7d1ac025c0606f5dff0907fd5ab3d55a94