[US_ME] Add edge case for validation (Recidiviz/recidiviz-data#34364)
## Description of the change

OVERDUE_FOR_DISCHARGE in ME will have 0 people eligible often. That's ok
and it's actually a number that we aspire for. So I added a where clause
in the validation so this doesn't pop up every time this happens.

## Related issues

Closes Recidiviz/recidiviz-data#34094

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0013a826e5cc237636ac604ba9bb4bddd84b77b8