Adds new span and events for INSIGHTS_USER unit of observation (Recidiviz/recidiviz-data#34300)
## Description of the change

Building off of Recidiviz/recidiviz-data#34299, this PR configures a new unit of observation for
`INSIGHTS_USER` and adds an associated span type
(`INSIGHTS_USER_REGISTRATION_SESSION`) that defines inclusion in the
population, as well as an events view that contains insights user
logins.

Note that the logins in that observations view are not restricted to
insights primary users, but when a logins metric is produced via agg
metrics that will join login events to the spans of time over which an
user is a primary registered Insights user according to insights primary
user registration sessions.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34100

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 52717876bad278ef482ed150b71fbf20b4eeff2e