[Insights][Actions] Add markdown preview on `ActionStrategyFormItem` (Recidiviz/recidiviz-data#34178)
## Description of the change

This adds in the markdown preview element to `ActionStrategyFormItem`.

After some investigation, it looks like this error is usually due to
conflicts/mismatches between versions of React and the package of
interest (in this case react-markdown):

```bash
ERROR in ./node_modules/react-markdown/lib/index.js 88:0-56
Module not found: Error: Can't resolve 'react/jsx-runtime' in '/Users/jen/dev/recidiviz-data/frontends/admin-panel/node_modules/react-markdown/lib'
Did you mean 'jsx-runtime.js'?
```

It turns out that starting with version 9.0.0, react-markdown requires
React 18, whereas we are on 17. This uses react-markdown 8.x.x instead.
I was never able to reproduce the error (I still don't understand why it
worked for me even though we should have all the same package versions)
so I'm not sure if this is the right fix, but let me know if it works!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34134

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8130fcf1ab2438f09317fc47bb8eedae9295b390