[Justice Counts] Computationally Fetch `available_metrics` and `available_sectors` in dashboard homepage API. (Recidiviz/recidiviz-data#34307)
## Description of the change

This PR adds logic so that we will fetch the available metrics and
sectors for agencies when visiting the agency dashboard landing page.

At first, we only agreed to implement available sectors and
is_dashboard_enabled:

```
available_sectors: A list of systems that the agency is reporting for. This list should only include sectors if the agency has enabled at least one metric in that sector with at least one datapoint.

is_dashboard_enabled: A required Boolean field for each agency that indicates whether an agency’s data can be rendered as a dashboard. If false, the agency’s data should not be rendered as a dashboard. If true, the following criteria are met: 1) CSG has enabled the is_dashboard_enabled field, and the public dashboard is active. 2) The agency has at least one metric enabled, with at least one data point available.
```

As I started to implement this, I realized that we should include
`available_metric_keys` in this version because we need that data to
return `available_sectors` and `is_dashboard_enabled` anyway.

```
available_metric_keys: A list of enabled and valid metric keys used across all JC metrics. To appear on this list, the agency must have enabled the metric and at least one datapoint for the metric.
```

As I was doing this! I realized it made sense to throw
`available_disaggregations` into the response as well, given that we
were looping through all the report datapoints for an agency anyway.

```
available_disaggregations: A dictionary of available metrics and the enabled disaggregations and dimensions. If an agency has enabled the ‘PRISONS_FUNDING’ metric, then this would be included in the dictionary as a field. Each enabled metric disaggregation key for funding would be included with a list of enabled dimension keys. For example:
```

[Here's](https://nichelletest---publisher-web-b47yvyxs3q-uc.a.run.app/api/agencies/dashboard)
what it looks like in staging!

At first I was getting `Service Unavailable 503` errors when visiting
the link on my playtesting branch, so I decided (with the suggestion of
Lili) to only return agencies with enabled dashboards via the admin
panel. After this, I still saw the error. When I checked the staging
logs I saw this error:
```
Memory limit of 2048 MiB exceeded with 2054 MiB used. Consider increasing the memory limit, see https://cloud.google.com/run/docs/configuring/memory-limits
```

Turns out that when we do a staging release on playtesting we give
limited memory to the release. I increased it to 2GiB and was able to
load the response.

## Related issues

Related to https://github.com/Recidiviz/recidiviz-data/issues/34211

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 88416cd7a305e25d5bd90ccc2b42d06eb3b9fbf9