[US_AZ] Refactor Home Plan preprocessed view (Recidiviz/recidiviz-data#34366)
## Description of the change

This PR makes some changes to the preprocessed Home Plan view in AZ to
make it more accurate.
- Makes it so that "Address Denied" statuses indicate that a home plan
is in progress rather than denied outright. We've learned that home
plans can be and often are resubmitted with a different address in this
scenario, so it an address denial doesn't mean that a person's home plan
is denied forever. An address denial can be a step on the way to
approval.
- Correctly saves only the most recent status of a given home plan (this
logic was buggy before)
- Collapses multiple CTEs that had little purpose into one that combines
their functionality
- Adds comments

**One thing in particular requires attention in review!** This view
attempts to connect home plans to incarceration stints in a one-to-one
relationship. I confirmed that the first CTE produces exactly one home
plan status per DOC_ID, which is in line with our goal. The
`status_joined_to_sessions` CTE is meant to provide the `session_id` of
each particular incarceration stint for easier identification; however,
it is producing multiple session_ids per home plan. This is likely the
result of buggy join logic to `compartment_sessions`. I need some help
correcting that, if possible! @shalindb @hugosr-r

See results in
`recidiviz-staging.elise_analyst_data.us_az_home_plan_preprocessed_materialized`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5fa6ea842b093aae29bfcb9cf993631485f1053f