[Workflows] Add is_surfaceable to person impact funnel (Recidiviz/recidiviz-data#34381)
## Description of the change

Add the `is_surfaceable` flag onto the Workflows person impact status
sessions to enable users to flag when a person and task type was/was not
exported to the front end dashboard in conjunction with the other impact
funnel status sessions flags.

This change indicates we currently have 55K surfaceable opportunities.

## Related issues

Closes Recidiviz/recidiviz-data#34380

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 28edc06839c74c58e342e8ca2e4158f75c5b3235