Adds is_fully_launched flag to launch metadata yaml (Recidiviz/recidiviz-data#34411)
## Description of the change

Following [this slack
thread](https://recidiviz.slack.com/archives/C07NX5PM0JG/p1729630827330269)
and corresponding update to All Launched Workflows google sheet to
include partial launches with a separate flag for whether or not an
opportunity has been fully launched, this PR updates the yaml file to
include the new `is_fully_launched` column in the schema for
`workflows_launch_metadata`, and pulls this field directly into the
completion event type metadata table rather than inferring that an
opportunity is fully launched if it appears in the sheet.

This will fix a bug that is currently relying on the sheet to only
include fully launched workflows, which means that some task types are
misrepresented in the workflows impact looker explore.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 5b349e1f2ecd133c27393fe8c0e2febdeaf968a1