[raw data import] Supply encoding and delimiter to BQ load job (Recidiviz/recidiviz-data#34441)
## Description of the change

supply encoding/delimiter so BQ valid encodings/delimiters that are not
utf-8/a comma are correctly parsed

In most cases we can upload ISO-8859-1 as UTF-8 to BQ without specifying
encoding, but there are some cases in which we would upload junk to BQ
if we attempted to upload an ISO-8859-1 file without specifying
ISO-8859-1 encoding. ex the NBSP character is encoded as `0xA0` in
ISO-8859-1 and ` 0xC2 0xA0` in UTF-8, and `0xA0` is not valid in UTF-8.
the test file in this PR is an ISO-8859-1 file containing NBSP
characters, which when uploaded as ISO-8859-1 can then be trimmed during
the transform.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34440

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a60f9b83fb1361c6ff0c87c36c201e2a405b0efa