[admin panel] always let flash checklist user click Mark Done (Recidiviz/recidiviz-data#34385)
## Description of the change

Always allows user to click `Mark Done` button if the `Action Button`
doesn't return.

**_[!!!!]_** is is important to note that it is a feature of this design
that clicking the `Mark Done` button is not supposed to cancel the
request of the `Action Button` -- you can see in the video below (where
i added a `sleep` in the endpoint), that the endpoint will eventually
return. if it does, it will also increment the step section twice
errantly, so i guess you sort of need to make sure that the endpoint is
not returning. adding abortcontrollers to all of the endpoints felt like
more work than i wanted to do for this use case

quick video demo:

https://github.com/user-attachments/assets/ee4b268d-b317-4183-88fb-863b51e2b178

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18624

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2bde86d656d9265480037db74475a879c6d00fed