[US_MI][Workflows] Update start dates for add and warden review counts (Recidiviz/recidiviz-data#34443)
## Description of the change

We got feedback from TT that they were having to backdate multiple
reviews per resident since the case note field for warden and add
reviews was not commonly used prior to launch. Here we update the date
we begin accumulating reviews to be 6 months or 1 year before the launch
date, so users only have to input the latest review per resident.

This results in 1/111 fewer residents eligible for the warden review,
and 9/45 fewer residents eligible for the add review.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
#[6513](https://github.com/Recidiviz/recidiviz-dashboards/issues/6513)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fa5b4988c55638bcdd98051156e6fbe849dc959a