[US_AZ] Pull flat sentence status into state sentence metadata (Recidiviz/recidiviz-data#34456)
## Description of the change

We found out yesterday that a sentence being a "flat sentence"
automatically disqualifies someone from DTP and TPR. Since this
information is relevant to the TPR tool we are building, this PR makes
it accessible downstream on the `state_sentence` entity.

This PR ingests the flat sentence status of _every_ sentence handed down
in AZ. In reality, we only want to know if the _ruling_ sentence of a
given sentence group is a flat sentence. When we use this data, we will
need to inner join `state_sentence` to `state_sentence_group` using
`sentence_group_id` so that we only access the flat sentence status
relating to the ruling sentence of a given group.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#34435

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c6bf0cc1d9f776d79817c6a5caea6a1ed5af00d4